move types to files
